separate cfn output irsa logic
